Remove dependency on 🤗 transformers
Before we this change, we would always load the model weights using the
`transformers` library, even when deserializing finetuned models. This change
postpones model loading to the `init` callback of the piece and transformer
encoder models. In this way, we only attempt to load models when a new pipeline
is initialized.

To further remove the dependency on the `transformers` library, we make the
actual loader configurable through the new `model_loaders` registry. This allows
us to provide other loaders in the future. This is an example configuration of a
transformer pipe using the model loader mechanism:

```
[components.transformer]
factory = "curated_transformer"

[components.transformer.model]
@architectures = "curated-transformers.XLMRTransformer.v1"
vocab_size = 250002

[components.transformer.model.encoder_loader]
@model_loaders = "curated-transformers.HFEncoderLoader.v1"
name = "xlm-roberta-base"

[components.transformer.model.piecer_loader]
@model_loaders = "curated-transformers.HFSentencepieceLoader.v1"
name = "xlm-roberta-base"

[components.transformer.model.with_spans]
@architectures = "curated-transformers.WithStridedSpans.v1"
stride = 96
window = 128
```

Since we cannot rely on the `transformers` library anymore to provide a model's
hyperparameters, we use the default hyperparameters for a model instead. The
configuration should specify deviations from the default configuration,
typically the vocab size, hidden size, number of layers, and number of attention
heads.