Refactor pipeline and listeners to allow the processing of all transformer outputs
Models that follow the transformer in the pipeline model chain operate on the outputs of all the transformer layers (if required by downstream components). This will allow for the usage of a separate listener that can pass a weighted sum of the former to the pooling layer (as opposed to using just the last layer's hidden state).