Move embedding projection to BertEmbeddings
This change moves the embedding projection from AlbertEncoder to
BertEmbeddings. The projection it activated when the embedding and
hidden sizes differ.

By moving the projection to the embedding layer, we can support
distallation with an embedding projection without fully switching to
the ALBERT model. Furthermore, the projection can be used to compress
embedding layers of existing BERT/RoBERTa models.