Speed up pooling by concatenating all transformer outputs
In our listeners, we pooled output per doc or per doc + layer. This
results in a lot of kernel (bookkeeping) overhead and provides limited
parallelization. This change introduces two wrappers for transformer
output pooling (`with_ragged_layers` and `with_ragged_last_layer`)
that concatenate all the `Ragged`s in transformer outputs to a single
`Ragged`. This allows us to pool the outputs in one go. In distillation
experiments with L2 layer loss, this reduced the amount of GPU time
that was used for pooling from ~48% to 8%.