Increase GPU throughput during gradient accumulation (#79)
* `Transformer`: Defer loss calculation until final backprop callback

Originally, we were updating the pipe's cumulative loss value for every layer being backpropagated into, for every listener. This resulted in significant overhead when using the `ScalarWeightingListener` as it led to a lot of consecutive kernel launches/teardowns, which was also exacerbated by the number of operations required for the SSE calculation.

Now, we merely accumulate the gradients from all listeners and calculate the loss once at the end of the accumulation phase. This improves the throughput of the `ScalarWeightingListener`'s backprop pass considerably.

* Add comment explaining the change in loss semantics