Add character encoder (#110)
This encoder splits a token into characters and looks up their
identifiers. The primary motivation for adding this encoder is that our
standard model for Japanese requires it. Along with the encoder itself,
`curated-transformers.CharEncoderLoader.v1` is added to load character
vocabularies and support to load Japanese character-based encoders is
added to `curated-transformers.HFPieceEncoderLoader.v1`.