Add optional `piece_encoder` argument to each transformer architecture
This allows a configuration to override the piece encoder that is used
by the architecture, which is needed in rare cases. For instance, the
`cl-tohoku/bert-base-japanese-char-v2` model uses a BERT transformer but
a character-based piecer.

By default the `piece_encoder` value is `None`, which will use the
default piece encoder for the architecture. This ensures that there is
no configuration overhead in the vast majority of cases.