Add support for Torch `scaled_dot_product_attention` (#128)
* Add support for Torch `scaled_dot_product_attention`

Torch now has a function for computing the scaled dot product attention,
`scaled_dot_product_attention`. This change adds support for using this
function in our self-attention layer. However, since this function is
still in beta and at not faster at the moment, we feature-gate it
through the `CURATED_TORCH_SDPA=1` environment variable.

* Fix shape comments

* Attempt to silence attr error on older Torch versions

* The ignores never go where they make sense