RotaryEmbeddings.forward: add support for a positions tensor
In most generation scenarios, we are only applying the rotary embeddings
to new (uncached) positions. So we need a way to specify what the
positions of the inputs are to apply the correct rotary embedding. We
cannot use a single offset, since different batch items may have
different offsets. For instance, when processing two prompts with
different lengths:

A B C D E
    X Y Z

The initial offsets are:

0 1 2 3 4
    0 1 2

So when generating the next token, we need the offsets [[5], [3]].