Improved Hugging Face Hub model loading infra, apply to RoBERTa (#153)
* Improved Hugging Face Hub model loading infra, apply to RoBERTa

This change is the first in a series of changes that moves HF model
loading from an idempotent function to the models themselves. The
idempotent function got quite convoluted when adding decoder
architectures, furthermore it probably makes more sense to couple
model-specifics to the model. This change introduces the
`FromPretrainedHFModel` mixin. Modules that want support for loading
parameters from Hugging Face Hub can derive from this mixin and
implement the `convert_hf_state_dict` and `from_hf_config` methods. The
mixin will then provide the `from_hf_hub` method to download a model
from the Hugging Face Hub.

This change also migrates the RoBERTa models to use the mixin.

* FromPretrainedHFModel: remove transformers dependency

Instead download the model configuration and checkpoint with
`huggingface_hub`, which is a rather small package.

* Docstring fixes

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>

* Typo

* Remove comment about max length

The rationale is already described below the comment.

---------

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>