Add GPT-NeoX decoder and causal language model
In order to support outputting predicted tokens and key/value caching,
two new output classes are added: `ModelOutputWithCache` and
`CausalLMOutputWithCache`. The existing `PyTorchTransformerOutput` is
renamed to `ModelOutput`.

This change also adds a new `SelfAttentionWithRotaryEmbeddings` layer,
which is similar to the existing self-attention layer except that it
also applies rotary embeddings.

This change also adds support for causal masking to the self-attention
mechanism to support autoregressive models.