Allow more fine-grained control over tokenizer input
Before this change, tokenizers took `List[str]` is their input. However,
this is not ideal in some cases. For instance in many LLMs we have to
provide a model with special pieces to signal e.g. the start/end of a
prompt. To serve these use cases, this change allows more fine-grained
control over the input.

Besides supporting `List[str]`, the tokenizers now also support
`List[InputChunks]`. `InputChunks` is a list-derived data structure
that can contain one of two types of chunks:

* `TextChunk`: this is a regular piece of text that should be tokenized.
* `SpecialPieceChunk`: this type of chunk stores a single piece that
  should be looked up in directly in the vocabulary. Additionally, it
  can have strings to be appended/prepended to the preceding/succeeding
  text chunks.

The tokenizer normalizes these chunks by merging consecutive text chunks
and adding the strings specified in special piece chunks. After this
merge step, it processes the chunks.

Since we can have special pieces in the input with this change,
begin/end of sequence markers can now be added safely in the
pre-encoders.

In order to test the use of special piece chunks in the input, the
GPT-NeoX tokenizer is updated to include special pieces in the
vocabulary. We may eventually want to do this for other types of
tokenizers as well.