Add a generator
This change adds a very basic generator. It only does argmax decoding,
no sampling, no beam search, no logit temperatures, no stopping
conditions other than the EOS piece, etc. The goal is to put something
in place that we can iterate on. The main classes added in this change
are:

- `Generator` provides the generation loop.
- `StringGenerator` is a wrapper for `Generator` that takes pieces
  and produces pieces.
- `DollyV2Generator` uses these two classes and wraps user prompts in
  model-specific instructions.

There are many other changes throughout the code. The largest is that we
introduce a generic type bound by `CacheProtocol`, so that we can treat
cache (such as the key-value cache) in a generic way.

This change also adds a `veryslow` test mark for tests that we probably
don't want to run in CI because the models are too large and/or slow,
such as the Dolly V2 generation test.