`StringGenerator`: decode all output ids at once (#185)
* StringGenerator: decode all output ids at once

Generator returns an iterator, allowing the piece identifiers to be
generated incrementally. Similarly, StringGenerator also returned an
iterator, converting the piece identifiers to pieces. However, the
string decoding used for this can fail since not every piece identifier
represents a valid UTF-8 string by itself. For example, the byte BPE
encoder uses bytes as the smallest unit, and a single byte can represent
a UTF-8 initial or continuation byte. Attempting to convert such a byte
to a string will be invalid.

Another issue is that potential decoding pre/post-processors may expect
to see the full sequence.

This change addresses this shortcoming by accumulating the piece
identifiers in StringGenerator and only decoding the piece identifiers
once all sequences are done generating.

We might want to bring back incremental string generation in the future,
but we need to extend the tokenizers to support this properly.

* Remove debugging prints