Add support for stop conditions
This change introduces a new base class, StopCondition. Implementations
of this class provide stop conditions for generation. A stop condition
gets the most recent parse state and updates one of two arrays that
mark generation of a sequence as completed:

- completed_exclude: for sequences that are completed.
- completed_include: for sequences that are completed and for which
  the last generated piece should still be emitted.

Initially, there are two stop conditions:

- EndOfSequenceCondition: stop generating a sequence when the
  end-of-sequence piece id is predicted. This replaces the previous
  manual handling of the end-of-sequence piece id.
- MaxNewPiecesCondition: stop generating when the maximum number
  of generation steps has been reached.

The generator and generator state has been adjusted to work with stop
conditions.