Add FalconGenerator and all the needed machinery
This change adds support for generation using the Falcon models. The
generator wrapper is FalconGenerator. Falcon uses the Refined Web Model,
which is slightly different from other models that we currently have:

- The model can use shared key and value heads. This is implemented
  using the new qkv_head_sharing argument of the self attention layer
  along with the QkvHeadSharing enum.
- The model can use all linear layers without bias. This is accomodated
  by adding a new use_bias argument to the self attention layer and
  the pointwise feed-forward layer.
- The activation is always GELU.

Unfortunately, we cannot test decoding with cache, position tensors,
or the causal LM against the upstream implementation, since it is
broken in two ways:

- The upstream implementation uses the PyTorch scaled dot product
  attention, which creates invalid causal attention masks. When
  the query and key shapes differ.
- The upstream implementation does not take into account that rotary
  embeddings should be offset for position when using a cache. As
  a result, the new query and key pieces will always be treated as
  starting at position 0.

I still need to investigate the differences between the NeoX tokenizer
and the Refined Web Model tokenizer. I'll address this in a future
PR.