Add support for the `tokenizers` package
This change adds support for the `tokenizers` package. We started
implementing support for reading `tokenizer.json` files. However, fully
supporting this format is currently not our focus, so instead we use the
`tokenizers` package to support these tokenizers.

To support this, the class hierarchy needs to be restructed a bit:

* The original `Tokenizer` class is split into two classes:
  `TokenizerBase` defines the interface for tokenizers. `LegacyTokenizer`
  implements the interface for our original tokenizers.
* The new `Tokenizer` class it the `tokenizers`-based tokenizer. This
  tokenizer has a prominent name, because going forward it will be
  the preferred/default tokenizer.