Replace decoder layers by a shared layer (#248)
* Replace decoder layers by a shared layer

This change replaces the decoder layers by a shared decoder layer,
`DecoderLayer`. Since the overlap between encoder and decoder layers is
very large, both are factored out to a single `_TransformerLayer`.

The `_TransformerLayer` class is not public because exposing its
`forward` method would violate the Liskov substitution principle, but it
holds the implementation of both types of layers.

The Falcon decoder layer does not currently use the shared decoder layer
because its dropout use is nonstandard. However, I might try to support
this as well in a follow-up PR.

This change also updates the Dolly V2 tests. The different order of
summing in the residual connection leads to different outputs due to the
imprecision of bfloat16.

* Docstring fixes

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>

---------

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>