Add support for TorchScript tracing
This change adds support for TorchScript tracing and tests tracing with
all of our models to ensure that there are no regressions.

Most of the technical details of this PR are in the developer docs,
but it boils down to: we can't use arbitrary data structures in
traced models. So we make non-Tensor arguments and return types
dicts as well (through `DataclassAsDict`) or convertable to tuples
(through `DataclassAsTuple`). We need a small amount of internal
handling to reinterpret argument dicts ash their counterpart types and
to return tuples from models.

This change also adds a new _Deployment_ section to the documentation
that describes how to trace models.