Make `DataclassAsTuple` types actual tuples
Before this change, `DataclassAsTuple` provided an `astuple` method to
convert the dataclass to a tuple. This change makes `DataClassAsTuple` a
subclass of `tuple`, so that it can be used directly in TorchScript
tracing.

This removes the need to conditionally return tuples depending on
whether we are tracing.

We also make `KeyValueCache` derive from `DataclassAsTuple` rather than
`DataclassAsDict`. This provides weaker guarantees in rewrapping (since)
we don't get keys anymore. However, the benefit is that we can run
tracing with strictness switched on (since the return value does not
contain a dict).