More renames in preparation for the release (#277)
* LLaMA -> Llama

The casing of LLaMA was changed with the introduction to Llama2. Since
it's easier to type and easier to remember this casing, let's use that
in curated as well.

* parallel_attention -> use_parallel_attention

* dim -> width

We decided a while ago to use _width_ for vector widths, since dim is
confusing (since it's also the length of the shape). Fix a few more
cases.

* Import fix