Register models using `catalogue` (#351)
* Register models using `catalogue`

So far we have hardcoded the available encoders/decoders/causal LMs in
the auto classes. This has the downside that the auto classes only work
with models that are provided by Curated Transformers.

This change adds registries for encoders/decoders/causal LMs. The auto
classes query the relevant registry and check which registered model
supports the downloaded model (through the `hf_model_types` method of
the `FromHFHub` mixin). This makes it possible to register external
models with Curated Transformers, so that they can also be used with the
auto classes.

Adding registries for tokenizers and generators is deferred to future
PRs.

* Improve warning

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>

* Registry docs: link to an example `from_hf_hub`

---------

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>