Add support for loading parameters in-place (#370)
In some applications (e.g. spaCy Curated Transformers), we may already
have constructed the model and we want to load parameters in-place. This
change adds in-place versions of the `from_*` class methods.

To support this properly, the in-place loaders should not need to access
the configuration anymore. So, the Torch dtype deserialization has moved
from the `from_repo` method to the configuration deserialization. To
support this, all model configurations now also take a `dtype`
parameter.